requirements.txt
asgiref==3.7.2
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.4.2
Django==5.0.4
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
idna==3.10
Markdown==3.8
PyJWT==2.9.0
PySocks==1.7.1
redis==6.0.0
requests==2.31.0
sqlparse==0.5.3
urllib3==2.4.0
gunicorn==20.1.0
